Fixed errors in SELECT/JOINs in trigger function
#101
The previous issue with it hanging was actually errors caused by the
query trying to select movement from miovision.intersections. This query
solves the problem by first selecting the data to an intermediate table.
The function also changes the trigger so that inserts data to volumes
after it has been inserted to raw_data.