Separated updating the indexes to a separate function
#101
Recently, I've had issues with updating the indexes. The run time seems
to take about an hour. To solve this issue, I splited that part of the
process to a separate function, and changed it to include a where clause
that searches within specific dates specified by the python script. Now,
the run time is below 20 seconds to update the indexes