Extreme value analysis for ATR permanent #161 #162
Almost done. Created process and did most of the QC.  I still need to finish the QC (for zero values that occur in an hour with other small values that are not zero), potentially reconfigure my way of identifying outliers slightly, and then create final open data ATR permanent view.