Made performance improvements and created views for gap filled views
The aggregate_speed_counts_15min has a WHERE clause for the speed bins
to avoid the same problems as the Miovision query.

For the gap filled volumes_15min view, it uses report_dates to get a
list of valid dates. volumes_average is a view of the average count for
each bin/day of week/month/sign location combination. volumes_15min_full
fills data from volumes_15min, and if there are gaps, uses data from the
averages view to fill it in.

#173