#234  updated bliptrack TT routes with new detectors
One detector at Bloor and Runnymede added in the 'Locations' sheet. In segments table, analysis ID 1413812 refers to Dovercourt Rd from College to Dundas. The analysis id refers to detector names AC1 and AD3. both of which are inactive since 2018 August. Detector AD2 also is inactive since Aug 2018. Located at Sterling and Dundas. Not included for route in the sheet. If that route is needed, I will add. CO_DF to AD2 (EB and WB)