#234 Updated with BT detector at Highpark
Green ones are high priority. Green DU_UN and BD_AV is along the Active TO bike lane. May be of interest. Dufferin (both direction) is a low priority.

Routes along Roncesvalles is red as BD_DU is not working. But might be a route of interest in future.

Highpark-Runnymede both directions and Highpark-Dufferin both directions are the four added routes of immediate interest.